requirements.txt
audioop-lts==0.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
lxml==6.0.2
pillow==12.0.0
pydub==0.25.1
pyTelegramBotAPI==4.29.1
python-docx==1.2.0
requests==2.32.5
SpeechRecognition==3.14.4
standard-aifc==3.13.0
standard-chunk==3.13.0
telebot==0.0.5
typing_extensions==4.15.0
urllib3==2.6.2
